Clean up pod proxy connection accounting
