+++ pkg/abstractions/pod/proxy.go
@@ -26,7 +26,6 @@ import (
 	"github.com/google/uuid"
 	"github.com/gorilla/websocket"
 	"github.com/labstack/echo/v4"
-	redis "github.com/redis/go-redis/v9"
 )
 
 const (
@@ -35,27 +34,7 @@ const (
 	connectionKeepAliveInterval   time.Duration = time.Second * 1
 	connectionReadTimeout         time.Duration = time.Minute * 5
 	containerAvailableTimeout     time.Duration = time.Second * 2
-	connectionSyncInterval        time.Duration = 100 * time.Millisecond
-	connectionSnapshotTTL         time.Duration = 2 * time.Second
-	reverseProxyBufferSize        int           = 32 * 1024
 )
-
-var reverseProxyBufferPool = sync.Pool{
-	New: func() any {
-		buf := make([]byte, reverseProxyBufferSize)
-		return &buf
-	},
-}
-
-type podReverseProxyBufferPool struct{}
-
-func (podReverseProxyBufferPool) Get() []byte {
-	return *reverseProxyBufferPool.Get().(*[]byte)
-}
-
-func (podReverseProxyBufferPool) Put(buf []byte) {
-	reverseProxyBufferPool.Put(&buf)
-}
 
 type container struct {
 	id              string
@@ -470,7 +449,7 @@ func (pb *PodProxyBuffer) backendProxy(targetHost string) *httputil.ReverseProxy
 	proxy := &httputil.ReverseProxy{
 		Director:   func(req *http.Request) {},
 		Transport:  pb.backendTransport(targetHost),
-		BufferPool: podReverseProxyBufferPool{},
+		BufferPool: abstractions.ProxyBufferPool{},
 		ErrorHandler: func(rw http.ResponseWriter, req *http.Request, err error) {
 			http.Error(rw, "Backend route unavailable", http.StatusBadGateway)
 		},
@@ -722,255 +701,6 @@ func (pb *PodProxyBuffer) checkContainerAvailable(containerAddress string) bool 
 	defer conn.Close()
 	metrics.RecordProxyBackendDialLatency("pod", pb.workspaceName(), pb.stubId, "discovery", true, time.Since(start))
 	return conn != nil
-}
-
-func (pb *PodProxyBuffer) incrementTotalConnections() (int64, error) {
-	val := pb.totalConnections.Add(1)
-	if val == 1 {
-		pb.publishTotalConnectionSnapshot(1)
-	}
-	return val, nil
-}
-
-func (pb *PodProxyBuffer) decrementTotalConnections() error {
-	decrementCounter(&pb.totalConnections)
-	return nil
-}
-
-func (pb *PodProxyBuffer) incrementContainerConnections(containerId string) error {
-	counter, ok := pb.containerConnectionCounter(containerId)
-	if !ok {
-		return nil
-	}
-	if counter.Add(1) == 1 {
-		pb.publishContainerConnectionSnapshot(containerId, 1)
-	}
-	return nil
-}
-
-func (pb *PodProxyBuffer) decrementContainerConnections(containerId string) error {
-	defer pb.signalWork()
-	go pb.setPodKeepWarmLock(containerId)
-
-	counter, ok := pb.containerConnectionCounter(containerId)
-	if !ok {
-		return nil
-	}
-
-	decrementCounter(counter)
-	return nil
-}
-
-func (pb *PodProxyBuffer) totalConnectionCount() int64 {
-	return pb.totalConnections.Load()
-}
-
-func (pb *PodProxyBuffer) containerConnectionCounter(containerId string) (*atomic.Int64, bool) {
-	if containerId == "" {
-		return nil, false
-	}
-	actual, _ := pb.containerConnections.LoadOrStore(containerId, &atomic.Int64{})
-	return actual.(*atomic.Int64), true
-}
-
-func (pb *PodProxyBuffer) containerConnectionCount(containerId string) int64 {
-	value, ok := pb.containerConnections.Load(containerId)
-	if !ok {
-		return 0
-	}
-	return value.(*atomic.Int64).Load()
-}
-
-func (pb *PodProxyBuffer) sharedContainerConnectionCount(containerId string) (int, error) {
-	if pb.workspace == nil {
-		return 0, nil
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
-	defer cancel()
-
-	connections, err := sharedPodContainerConnections(ctx, pb.rdb, pb.workspace.Name, pb.stubId, containerId)
-	if err != nil {
-		return 0, err
-	}
-	return int(connections), nil
-}
-
-func (pb *PodProxyBuffer) syncConnectionState() {
-	ticker := time.NewTicker(connectionSyncInterval)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case <-pb.ctx.Done():
-			return
-		case <-ticker.C:
-			pb.flushConnectionState()
-		}
-	}
-}
-
-func (pb *PodProxyBuffer) flushConnectionState() {
-	if pb.rdb == nil || pb.workspace == nil || pb.stubId == "" || pb.proxyId == "" {
-		return
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
-	defer cancel()
-
-	pipe := pb.rdb.Pipeline()
-	pipe.Set(ctx, Keys.podProxyConnections(pb.workspace.Name, pb.stubId, pb.proxyId, "total"), pb.totalConnections.Load(), connectionSnapshotTTL)
-	pb.containerConnections.Range(func(key, value any) bool {
-		containerId, ok := key.(string)
-		if !ok {
-			return true
-		}
-		count := value.(*atomic.Int64).Load()
-		if count == 0 {
-			pb.containerConnections.Delete(key)
-		}
-		pipe.Set(ctx, Keys.podProxyConnections(pb.workspace.Name, pb.stubId, pb.proxyId, containerId), count, connectionSnapshotTTL)
-		return true
-	})
-	if _, err := pipe.Exec(ctx); err != nil {
-		log.Debug().Err(err).Str("stub_id", pb.stubId).Msg("failed to publish pod connection snapshot")
-		return
-	}
-
-	pb.refreshConnectionAggregates(ctx)
-}
-
-func (pb *PodProxyBuffer) publishContainerConnectionSnapshot(containerId string, count int64) {
-	if pb.rdb == nil || pb.workspace == nil || pb.stubId == "" || pb.proxyId == "" || containerId == "" {
-		return
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
-	defer cancel()
-
-	if err := pb.rdb.Set(ctx, Keys.podProxyConnections(pb.workspace.Name, pb.stubId, pb.proxyId, containerId), count, connectionSnapshotTTL).Err(); err != nil {
-		log.Debug().Err(err).Str("stub_id", pb.stubId).Str("container_id", containerId).Msg("failed to publish pod container busy snapshot")
-	}
-}
-
-func (pb *PodProxyBuffer) publishTotalConnectionSnapshot(count int64) {
-	if pb.rdb == nil || pb.workspace == nil || pb.stubId == "" || pb.proxyId == "" {
-		return
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
-	defer cancel()
-
-	if err := pb.rdb.Set(ctx, Keys.podProxyConnections(pb.workspace.Name, pb.stubId, pb.proxyId, "total"), count, connectionSnapshotTTL).Err(); err != nil {
-		log.Debug().Err(err).Str("stub_id", pb.stubId).Msg("failed to publish pod total busy snapshot")
-	}
-}
-
-func (pb *PodProxyBuffer) refreshConnectionAggregates(ctx context.Context) {
-	total, _, err := sumRedisIntKeys(ctx, pb.rdb, Keys.podProxyConnectionsScan(pb.workspace.Name, pb.stubId, "total"))
-	if err != nil {
-		log.Debug().Err(err).Str("stub_id", pb.stubId).Msg("failed to sum pod total connections")
-		return
-	}
-
-	pipe := pb.rdb.Pipeline()
-	pipe.Set(ctx, Keys.podTotalConnections(pb.workspace.Name, pb.stubId), total, podContainerConnectionTimeout)
-	for _, c := range pb.availableContainerSnapshot() {
-		containerTotal, _, err := sumRedisIntKeys(ctx, pb.rdb, Keys.podProxyConnectionsScan(pb.workspace.Name, pb.stubId, c.id))
-		if err != nil {
-			log.Debug().Err(err).Str("stub_id", pb.stubId).Str("container_id", c.id).Msg("failed to sum pod container connections")
-			continue
-		}
-		pipe.Set(ctx, Keys.podContainerConnections(pb.workspace.Name, pb.stubId, c.id), containerTotal, podContainerConnectionTimeout)
-	}
-	if _, err := pipe.Exec(ctx); err != nil {
-		log.Debug().Err(err).Str("stub_id", pb.stubId).Msg("failed to publish pod connection aggregates")
-	}
-}
-
-func (pb *PodProxyBuffer) setPodKeepWarmLock(containerID string) {
-	if pb.containerRepo == nil || pb.workspace == nil {
-		return
-	}
-
-	if containerID == "" || pb.stubConfig == nil || pb.stubConfig.KeepWarmSeconds <= 0 {
-		return
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
-	defer cancel()
-
-	setPodKeepWarmLock(ctx, pb.containerRepo, pb.workspace.Name, pb.stubId, containerID, pb.stubConfig.KeepWarmSeconds)
-}
-
-func decrementCounter(counter *atomic.Int64) {
-	for {
-		current := counter.Load()
-		if current <= 0 {
-			return
-		}
-		if counter.CompareAndSwap(current, current-1) {
-			return
-		}
-	}
-}
-
-func sharedPodTotalConnections(ctx context.Context, rdb *common.RedisClient, workspaceName, stubId string) (int64, error) {
-	if rdb == nil || workspaceName == "" || stubId == "" {
-		return 0, nil
-	}
-
-	total, found, err := sumRedisIntKeys(ctx, rdb, Keys.podProxyConnectionsScan(workspaceName, stubId, "total"))
-	if err != nil || found {
-		return total, err
-	}
-
-	total, err = rdb.Get(ctx, Keys.podTotalConnections(workspaceName, stubId)).Int64()
-	if err == redis.Nil {
-		return 0, nil
-	}
-	return total, err
-}
-
-func sharedPodContainerConnections(ctx context.Context, rdb *common.RedisClient, workspaceName, stubId, containerId string) (int64, error) {
-	if rdb == nil || workspaceName == "" || stubId == "" || containerId == "" {
-		return 0, nil
-	}
-
-	total, found, err := sumRedisIntKeys(ctx, rdb, Keys.podProxyConnectionsScan(workspaceName, stubId, containerId))
-	if err != nil || found {
-		return total, err
-	}
-
-	total, err = rdb.Get(ctx, Keys.podContainerConnections(workspaceName, stubId, containerId)).Int64()
-	if err == redis.Nil {
-		return 0, nil
-	}
-	return total, err
-}
-
-func sumRedisIntKeys(ctx context.Context, rdb *common.RedisClient, pattern string) (int64, bool, error) {
-	var total int64
-
-	keys, err := rdb.Scan(ctx, pattern)
-	if err != nil {
-		return 0, false, err
-	}
-
-	for _, key := range keys {
-		value, err := rdb.Get(ctx, key).Int64()
-		if err == redis.Nil {
-			continue
-		}
-		if err != nil {
-			return 0, false, err
-		}
-		if value > 0 {
-			total += value
-		}
-	}
-
-	return total, len(keys) > 0, nil
 }
 
 func (pb *PodProxyBuffer) workspaceName() string {
